==> 08. Generating first-order datasets.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime
from config import SPLIT_CONFIG
from sklearn.experimental import enable_iterative_imputer 
from sklearn.impute import IterativeImputer

stations = ['Centar', 'Karpos', 'Lisice', 'Rektorat', 'Miladinovci']

df_extras = pd.read_csv('./data/extras', index_col = 0)
df_extras.head()

for station in stations:
    df_sensor = pd.read_csv(f'./data/truncated-sensor/{station}', index_col=0)
    df_weather = pd.read_csv(f'./data/truncated-weather/{station}', index_col=0)
    df = df_sensor.join(df_extras)
    df = df.join(df_weather)
    df.index = pd.to_datetime(df.index)
    df.to_csv(f'./data/first-order/{station}', index=True)
